+++ iphone/Classes/TiAppiOSProxy.m
@@ -30,7 +30,7 @@
     return @"Ti.App.iOS";
 }
 
--(void)_listenerAdded:(NSString*)type count:(int)count
+-(void)_listenerAdded:(NSString*)type count:(NSInteger)count
 {
 	if (count == 1 && [type isEqual:@"notification"]) {
 		[[NSNotificationCenter defaultCenter] addObserver:self selector:@selector(didReceiveLocalNotification:) name:kTiLocalNotification object:nil];
@@ -85,7 +85,7 @@
 
 }
 
--(void)_listenerRemoved:(NSString*)type count:(int)count
+-(void)_listenerRemoved:(NSString*)type count:(NSInteger)count
 {
 	if (count == 0 && [type isEqual:@"notification"]) {
 		[[NSNotificationCenter defaultCenter] removeObserver:self name:kTiLocalNotification object:nil];
