+++ iphone/Classes/TiUILabel.h
@@ -24,7 +24,7 @@
 -(TiLabel*)label;
 -(void)setPadding:(UIEdgeInsets)inset;
 -(void)setReusing:(BOOL)value;
--(int)characterIndexAtPoint:(CGPoint)p;
+-(NSInteger)characterIndexAtPoint:(CGPoint)p;
 @end
 
 #endif
