+++ iphone/Classes/TiUIListItemProxy.m
@@ -478,7 +478,7 @@ static void SetEventOverrideDelegateRecursive(NSArray *children, id<TiViewEventO
     NSArray* subproxies = [self children];
     NSInteger index=child?[subproxies indexOfObject:child]:-1;
     if(!child || NSNotFound != index) {
-        for (int i = index + 1; i < [subproxies count] ; i++) {
+        for (NSUInteger i = index + 1; i < [subproxies count] ; i++) {
             TiProxy* obj = [subproxies objectAtIndex:i];
             if ([obj isKindOfClass:theClass] && [obj canBeNextResponder]) {
 //                [[_listViewProxy tableView] scrollToRowAtIndexPath:_indexPath atScrollPosition:UITableViewScrollPositionNone animated:YES];
