+++ iphone/Classes/TiUIPicker.h
@@ -13,7 +13,7 @@
 @interface TiUIPicker : TiUIView<UIPickerViewDelegate, UIPickerViewDataSource> {
 @private
 	UIControl *picker;
-	int type;
+	NSInteger type;
 	
 	BOOL propertiesConfigured; // We're order-dependent on type being configured first, so have to re-configure after the initial setup.  What a pain!
 }
